TOPCOM patch
This is a hack... really, TOPCOM should be fixed to work properly for all cases. We should move away from the hack ASAP, hence the warning reminding users.

Conceptually, it'd be nicer to check if len(self.polytope().points())==3 (i.e., if the triangulation is of a polytope with no interior points). We can't do this, however, since TOPCOM still hangs on Triangulation([[0,0],[1,0],[0,1]]).neighbor_triangulations(), despite this polytope having 6 points.